Refactor code to manage Security Checkup row items in dedicated components
In addition to the above, I pushed all the state/querying dependencies into the components themselves.

The largest change in this area was for Two Factor Authentication, where we're now accessing user settings in a cleaner fashion.

I also added a check for Two Factor backup code verification that is only performed when 2FA is enabled.